// frontend/src/generated/models/ChartComponent.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Layout Dashboard API
 * API for managing dynamic dashboard layouts and components
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { ChartDataset } from './ChartDataset';
import {
    ChartDatasetFromJSON,
    ChartDatasetFromJSONTyped,
    ChartDatasetToJSON,
} from './ChartDataset';
import type { UIComponent } from './UIComponent';
import {
    UIComponentFromJSON,
    UIComponentFromJSONTyped,
    UIComponentToJSON,
} from './UIComponent';

/**
 * 
 * @export
 * @interface ChartComponent
 */
export interface ChartComponent extends UIComponent {
    /**
     * 
     * @type {string}
     * @memberof ChartComponent
     */
    chartType: string;
    /**
     * 
     * @type {ChartDataset}
     * @memberof ChartComponent
     */
    data: ChartDataset;
    /**
     * 
     * @type {string}
     * @memberof ChartComponent
     */
    title?: string;
}



/**
 * Check if a given object implements the ChartComponent interface.
 */
export function instanceOfChartComponent(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "chartType" in value;
    isInstance = isInstance && "data" in value;

    return isInstance;
}

export function ChartComponentFromJSON(json: any): ChartComponent {
    return ChartComponentFromJSONTyped(json, false);
}

export function ChartComponentFromJSONTyped(json: any, ignoreDiscriminator: boolean): ChartComponent {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        ...UIComponentFromJSONTyped(json, ignoreDiscriminator),
        'chartType': json['chartType'],
        'data': ChartDatasetFromJSON(json['data']),
        'title': !exists(json, 'title') ? undefined : json['title'],
    };
}

export function ChartComponentToJSON(value?: ChartComponent | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        ...UIComponentToJSON(value),
        'chartType': value.chartType,
        'data': ChartDatasetToJSON(value.data),
        'title': value.title,
    };
}