// frontend/src/generated/models/ChartDataPoint.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Layout Dashboard API
 * API for managing dynamic dashboard layouts and components
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ChartDataPoint
 */
export interface ChartDataPoint {
    /**
     * 
     * @type {string}
     * @memberof ChartDataPoint
     */
    label: string;
    /**
     * 
     * @type {number}
     * @memberof ChartDataPoint
     */
    value: number;
    /**
     * 
     * @type {string}
     * @memberof ChartDataPoint
     */
    unit: string;
}

/**
 * Check if a given object implements the ChartDataPoint interface.
 */
export function instanceOfChartDataPoint(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "label" in value;
    isInstance = isInstance && "value" in value;
    isInstance = isInstance && "unit" in value;

    return isInstance;
}

export function ChartDataPointFromJSON(json: any): ChartDataPoint {
    return ChartDataPointFromJSONTyped(json, false);
}

export function ChartDataPointFromJSONTyped(json: any, ignoreDiscriminator: boolean): ChartDataPoint {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'label': json['label'],
        'value': json['value'],
        'unit': json['unit'],
    };
}

export function ChartDataPointToJSON(value?: ChartDataPoint | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'label': value.label,
        'value': value.value,
        'unit': value.unit,
    };
}